Fixed the missing fields.
